Support Unicode paths on Windows
* All `char*` are internally UTF-8.
* All Win32 API calls presented with a UTF-8 path are
  converted to Unicode and now call the Unicode variant
  Win32 API.
* Where possible, the Unicode UNC prefix is prepended to
  paths to enable 64k maximum path length. Windows 10 has
  the ability for this maximum to be extended through a
  registry key.

NOTE: The FindFirstFile API call, with a Unicode UNC,
contains a character used as a wild-card. Because of this
fact, these paths are NOT prepended with the Unicode UNC
prefix. I could NOT find a way to escape the character, so
the API call would succeed. If anyone knows this answer,
please contact me!