Ensure WSL symlinks are skipped
With Windows 10, Microsoft introduced the Windows Subsystem
for Linux, with Ubuntu Linux running inside. As part of this
subsystem, Microsoft introduces a new reparse point code
which does not map to any other part of Windows, presently.
This reparse point was deemed required to handle symlinks
pointing to "special" locations that a normal Windows
symlink would not be able to represent.

For the time being, this new WSL symlinks are skipped; just
as system files are.