Improve read-only file transfers on Windows
Further improve upon transfers whereby the destination file
to be updated is read-only, on Windows. Read-only files are
unable to be opened for write-access, to update them, and
must have the read-only permission stripped from them prior
to being opened for writes through the hashing update
mechanisms Rsync utilizes to detect when a file indeed
needs a block updated within. Due to the fusion of calculating,
verifying, and possibly updating blocks; the caller must
ensure write-access is possible, when needed.